+++ web-scraper/data/ngdata/model.go
