+++ web-scraper/data/oda_backup/scraper/structs.go
